+++ scenarioo-client/app/app.js
@@ -84,3 +84,5 @@ require('./sketcher');
 require('./step');
 require('./useCase/usecase.controller.js');
 require('./diffViewer');
+
+angular.bootstrap(document, ['scenarioo']);